+++ application/tests/_ci_phpunit_test/functions.php
@@ -63,7 +63,7 @@ function reset_instance()
 	CIPHPUnitTest::loadLoader();
 
 	// Remove CodeIgniter instance
-	$CI = null;
+	$CI = new CIPHPUnitTestNullCodeIgniter();
 }
 
 /**