+++ docs/ChangeLog.md
@@ -2,6 +2,14 @@
 
 ## v0.10.0 (Not Released)
 
+### Fixed
+
+* Fix wrong implementation of resetting CodeIgniter instance. Now `reset_instance()` sets `null` to *CodeIgniter instance*. See [#74](https://github.com/kenjis/ci-phpunit-test/pull/74).
+
+### Changed
+
+* Now `$this->getDouble()` does not call the original constructor. See [#74](https://github.com/kenjis/ci-phpunit-test/pull/74).
+
 ### Added
 
 * NetBeans test suite provider `application/tests/_ci_phpunit_test/TestSuiteProvider.php`. To use it, go to *Project Properties* > *Testing*, check *Use Custom Test Suite* checkbox, and select the file.